Improve the way we compute volto styles
